+++ src/components/carousel/nav-tab-carousel/components/styled-components.js
@@ -7,7 +7,7 @@ var _interopRequireWildcard = require("@babel/runtime/helpers/interopRequireWild
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.StyledNextText = exports.StyledPrevText = exports.StyledArrowIcon = exports.StyledLeftArrowContainer = exports.StyledRightArrowContainer = void 0;
+exports.StyledNextText = exports.StyledPrevText = exports.StyledArrowIcon = exports.StyledLeftArrowContainer = exports.StyledRightArrowContainer = exports.LeftEmptyContainer = exports.RightEmptyContainer = void 0;
 
 var _styledComponents = _interopRequireWildcard(require("styled-components"));
 
@@ -25,36 +25,52 @@ var sharedHoverStyledArrowIcons = (0, _styledComponents.css)(["cursor:pointer;co
   var hoverColor = _ref2.hoverColor;
   return hoverColor;
 });
+var rightContainer = (0, _styledComponents.css)(["justify-content:flex-end;text-align:right;"]);
+var leftContainer = (0, _styledComponents.css)(["justify-content:flex-start;"]);
+
+var RightEmptyContainer = _styledComponents.default.div.withConfig({
+  displayName: "styled-components__RightEmptyContainer",
+  componentId: "jws9qo-0"
+})(["", " flex:1;"], rightContainer);
+
+exports.RightEmptyContainer = RightEmptyContainer;
+
+var LeftEmptyContainer = _styledComponents.default.div.withConfig({
+  displayName: "styled-components__LeftEmptyContainer",
+  componentId: "jws9qo-1"
+})(["", " flex:1;"], leftContainer);
+
+exports.LeftEmptyContainer = LeftEmptyContainer;
 
 var StyledRightArrowContainer = _styledComponents.default.div.withConfig({
   displayName: "styled-components__StyledRightArrowContainer",
-  componentId: "jws9qo-0"
-})(["justify-content:flex-end;text-align:right;", " &:hover{transform:translate(", ");", "}"], sharedStyledArrowIcons, (0, _typography.pxToRem)(8), sharedHoverStyledArrowIcons);
+  componentId: "jws9qo-2"
+})(["", " ", " &:hover{transform:translate(", ");", "}"], rightContainer, sharedStyledArrowIcons, (0, _typography.pxToRem)(8), sharedHoverStyledArrowIcons);
 
 exports.StyledRightArrowContainer = StyledRightArrowContainer;
 
 var StyledLeftArrowContainer = _styledComponents.default.div.withConfig({
   displayName: "styled-components__StyledLeftArrowContainer",
-  componentId: "jws9qo-1"
-})(["justify-content:flex-start;", " &:hover{transform:translate(-", ");", "}"], sharedStyledArrowIcons, (0, _typography.pxToRem)(8), sharedHoverStyledArrowIcons);
+  componentId: "jws9qo-3"
+})(["", " ", " &:hover{transform:translate(-", ");", "}"], leftContainer, sharedStyledArrowIcons, (0, _typography.pxToRem)(8), sharedHoverStyledArrowIcons);
 
 exports.StyledLeftArrowContainer = StyledLeftArrowContainer;
 var StyledArrowIcon = (0, _styledComponents.default)(_arrowIcon.ArrowIcon).withConfig({
   displayName: "styled-components__StyledArrowIcon",
-  componentId: "jws9qo-2"
+  componentId: "jws9qo-4"
 })(["min-width:", ";min-height:", ";"], (0, _typography.pxToRem)(9), (0, _typography.pxToRem)(9));
 exports.StyledArrowIcon = StyledArrowIcon;
 
 var StyledPrevText = _styledComponents.default.div.withConfig({
   displayName: "styled-components__StyledPrevText",
-  componentId: "jws9qo-3"
+  componentId: "jws9qo-5"
 })(["margin-left:", ";"], (0, _typography.pxToRem)(6));
 
 exports.StyledPrevText = StyledPrevText;
 
 var StyledNextText = _styledComponents.default.div.withConfig({
   displayName: "styled-components__StyledNextText",
-  componentId: "jws9qo-4"
+  componentId: "jws9qo-6"
 })(["margin-right:", ";"], (0, _typography.pxToRem)(6));
 
 exports.StyledNextText = StyledNextText;